Fixed formatting errors and refactord custom attributes to new namespace for KCC
