+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Character/Attributes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a0f84d4033713a4982676006a283e14
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
