+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Character/Attributes/ApplyGravityAttribute.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 72d1c9a90d98a1844af95e80e8dd3f78
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
