+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Character/Attributes/MovementSettingsAttribute.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c571a4d24d0012e47afb8d30dbe11718
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
